week2/yolov1/requirements.txt
matplotlib
numpy
pytorch
pandas
pillow
torchvision
tqdm